+++ src/pyquickhelper/helpgen/sphinx_main.py
@@ -58,6 +58,7 @@ from ..sphinxext.sphinx_blog_extension import BlogPostDirective, BlogPostDirecti
 from ..sphinxext.sphinx_runpython_extension import RunPythonDirective
 from ..sphinxext.sphinx_postcontents_extension import PostContentsDirective
 from ..sphinxext.sphinx_tocdelay_extension import TocDelayDirective
+from ..sphinxext.sphinx_youtube_extension import YoutubeDirective
 from ..sphinxext.sphinx_sharenet_extension import ShareNetDirective, sharenet_role
 from ..sphinxext.sphinx_template_extension import tpl_role
 from ..sphinxext.sphinx_epkg_extension import epkg_role
@@ -345,6 +346,7 @@ def generate_help_sphinx(project_var_name, clean=False, root=".",
     directives.register_directive("cmdref", CmdRef)
     directives.register_directive("postcontents", PostContentsDirective)
     directives.register_directive("tocdelay", TocDelayDirective)
+    directives.register_directive("youtube", YoutubeDirective)
     roles.register_canonical_role("sharenet", sharenet_role)
     roles.register_canonical_role("bigger", bigger_role)
     roles.register_canonical_role("githublink", githublink_role)