+++ src/pyquickhelper/sphinxext/__init__.py
@@ -24,6 +24,7 @@ from .sphinx_mathdef_extension import MathDef, MathDefList
 from .sphinx_nbref_extension import NbRef, NbRefList
 from .sphinx_postcontents_extension import PostContentsDirective, postcontents_node
 from .sphinx_tocdelay_extension import TocDelayDirective, tocdelay_node
+from .sphinx_youtube_extension import YoutubeDirective, youtube_node
 from .sphinx_sharenet_extension import ShareNetDirective, sharenet_node
 from .sphinx_template_extension import tpl_node
 from .sphinx_todoext_extension import TodoExt, TodoExtList
@@ -50,6 +51,7 @@ from ..sphinxext.sphinx_mathdef_extension import setup as setup_mathdef
 from ..sphinxext.sphinx_nbref_extension import setup as setup_nbref
 from ..sphinxext.sphinx_postcontents_extension import setup as setup_postcontents
 from ..sphinxext.sphinx_tocdelay_extension import setup as setup_tocdelay
+from ..sphinxext.sphinx_youtube_extension import setup as setup_youtube
 from ..sphinxext.sphinx_rst_builder import setup as setup_rst
 from ..sphinxext.sphinx_sharenet_extension import setup as setup_sharenet
 from ..sphinxext.sphinx_template_extension import setup as setup_tpl
@@ -101,7 +103,7 @@ def get_default_extensions():
                       setup_faqref, setup_exref, setup_nbref,
                       setup_docassert, setup_signature, setup_tpl,
                       setup_cmdref, setup_epkg, setup_rst, setup_postcontents,
-                      setup_tocdelay, setup_releases,
+                      setup_tocdelay, setup_youtube, setup_releases,
                       # directives from sphinx
                       setup_graphviz, setup_math, setup_todo,
                       # the rest of it